requirements.txt
aiohttp
beautifulsoup4
python-dotenv
supabase